Integration tests refactors
