+++ src/test/java/integration/com/project/trackfit/PersonalTrainerControllerIntegrationTest.java
@@ -1,34 +1,42 @@
-package com.project.trackfit.trainer;
+package integration.com.project.trackfit;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.project.trackfit.Main;
 import com.project.trackfit.core.APICustomResponse;
 import com.project.trackfit.trainer.dto.PersonalTrainer;
 import com.project.trackfit.trainer.repository.PersonalTrainerRepository;
 import com.project.trackfit.user.dto.ApplicationUser;
 import com.project.trackfit.user.repository.ApplicationUserRepo;
 import com.project.trackfit.user.component.Role;
-import com.project.trackfit.utils.AbstractIntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import java.util.Collections;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @ExtendWith(SpringExtension.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-class PersonalTrainerControllerIntegrationTest extends AbstractIntegrationTest {
+@SpringBootTest(classes = Main.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class PersonalTrainerControllerIntegrationTest {
+
+    @LocalServerPort
+    private int port;
 
     @Autowired
     private PersonalTrainerRepository personalTrainerRepository;
@@ -36,12 +44,17 @@ class PersonalTrainerControllerIntegrationTest extends AbstractIntegrationTest {
     @Autowired
     private ApplicationUserRepo applicationUserRepo;
 
+    @Autowired
+    private TestRestTemplate restTemplate;
+
     private UUID trainerId;
+    private HttpHeaders headers;
 
     @BeforeEach
     public void setUp() throws JsonProcessingException {
-        super.abstractSetUp();
-
+        headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.setAccept(Collections.singletonList(MediaType.APPLICATION_JSON));
         trainerId = UUID.randomUUID();
 
         String email = "email@example.com";
@@ -87,5 +100,24 @@ class PersonalTrainerControllerIntegrationTest extends AbstractIntegrationTest {
 
         PersonalTrainer savedTrainer = personalTrainerRepository.save(trainer);
         return savedTrainer.getId();
+    }
+
+    protected void registerUser(String email, String firstName, String lastName, Integer password, Role role, Integer age, String address) {
+        String registerUrl = "http://localhost:" + port + "/api/v1/auth/register";
+        HttpEntity<String> registerRequest = new HttpEntity<>(String.format("{\"email\":\"%s\",\"firstName\":\"%s\",\"lastName\":\"%s\",\"password\":\"%s\",\"role\":\"%s\",\"age\":\"%s\",\"address\":\"%s\"}", email, firstName, lastName, password, role.name(), age, address), headers);
+        restTemplate.postForEntity(registerUrl, registerRequest, String.class);
+    }
+
+    protected String getJwtToken(String email, Integer password) throws JsonProcessingException {
+        String tokenUrl = "http://localhost:" + port + "/api/v1/auth/token";
+        HttpEntity<String> tokenRequest = new HttpEntity<>(String.format("{\"email\":\"%s\",\"password\":\"%s\"}", email, password), headers);
+        ResponseEntity<String> tokenResponse = restTemplate.postForEntity(tokenUrl, tokenRequest, String.class);
+
+        System.out.println("Response body: " + tokenResponse.getBody());
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        JsonNode jsonNode = objectMapper.readTree(tokenResponse.getBody());
+        System.out.println(jsonNode);
+        return jsonNode.get("data").get("accessToken").asText();
     }
 }
